When checking for a qualifier on an element, check to make sure the element is superficially valid.
If it's not, the qualifier (or the type) may be generated, and therefore we need to defer processing.

In addition to regular bugs fixed in thew newly-added tests, this had an interesting fix for AOT. If the qualifier wasn't yet-generated when generating a subcomponent implementation, but was generated by the time an ancestor reimplemented the child, we could try to implement the missing binding method for the child (unqualified key) with a modifiable method for the qualified key. Those method names wouldn't align (based on how we name methods) and javac would complain in turn.

RELNOTES=Properly generate components when qualifiers are generated

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227589737